lerobot-web-be/requirements.txt
fastapi[standard]
uvicorn
starlette
aiortc
python-multipart
opencv-python
transformers
pytest
feetech-servo-sdk
placo
pydantic_settings